Update patch set 2
Patch Set 2: Verified

Was installing vs2013, which happens to break MSBUILD (and thus vs2010). So, I tried using my mingw shell... the build was broken. 

It works with this patch, but it's still a bit noisy: Mingw g++ v4.5.2 is angry about a bunch of the printf()s in sample.cpp and vttreader.cpp.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=+1