Update patch set 1
Patch Set 1:

I can replace makefiles and unix with vcxproj and windows-- the point is that it has to be handled somewhere-- here or by downstream projects. Either way, like I said, this is fine. I'll probably take care of adding a rename for windows later on today or tomorrow.

Patch-set: 1