Update patch set 5
Patch Set 5:

(3 comments)

Not sure where in the style guide it says wrapped statements start at the left most column except when in a return... I guess if we were using something like

return (statement);

That would make sense since it looks like our long if's. Anyway, not sure why clang-format is wrapping returns this way-- maybe it's just that I haven't seen this that much. Fine to deal with in a formatting specific patch.

Patch-set: 5