Update patch set 2
Patch Set 2:

>>> what happens (with the last frame timestamp tracking) if the
>>> first frame is a non-keyframe?
>> it's just going to write 0. for all practical purposes the first
>> frame cannot not be a key-frame (since we don't allow
>> non-monotonically increasing timestamps). should we throw an error
>> in that case?
> I'm not sure how the timestamp figures into this. I just wanted to > be sure that if someone was doing a stream copy there wouldn't be a > chance for bad data or accesses.

Timestamp figures into this because we write the timestamp of the last frame in the same track as ReferenceBlock for non-keyframes.

Since we only allow increasing timestamps, a non-keyframe with timestamp 0.1 cannot occur before a keyframe with timestamp 0. So, the first frame for each track will always be a keyframe and hence last_track_timestamp will always be set to a correct value (unless the input source itself is invalid).

Patch-set: 2