Update patch set 2
Patch Set 2:

> I don't see why you wouldn't want
> to have a full webm implementation in libwebm.

we certainly are not against having a full webm implementation in libwebm. you may have written it today but it's (mostly) on us to maintain this in the future. hence i don't want to include an API that will never be used just for the sake of completeness. :)

anyway, have you verified if this works as intended? an easy way to check would be to change sample_muxer to include a -copy_tags option or something equivalent.

Patch-set: 2