Update patch set 6
Patch Set 6:

(6 comments)

As Vignesh already touched on, the major missing piece here is addition of Tag related code to the sample. Please add usage of the tag code in sample.cpp and sample_muxer.cpp-- Since you're already using the tagging support in your own code, hopefully it's just a simple matter of a bit of copying/pasting and then adding a command line flag or two.

After adding sample support, please run the samples on files with tags within valgrind to check for memory leaks and any other issues. (or, if on a platform where this isn't really an option, let me know that's the case).

Patch-set: 6
Label: Code-Review=0