Update patch set 4
Patch Set 4:

(1 comment)

PTAL-- this is noisy and tedious, but requires attention to detail. Don't blame me: none of the magic numbers should have ever existed, and ReadUInt should _never_ have been used for ID reads-- the fact that it was is either laziness or insanity (pick one or both!). This is a _shit show_, or rather, it's the beginnings of cleaning one up.

Anyway, we now have sane ID usage in mkvparser (\o/ ?). This works on my machine, but I have not had a chance to test it exhaustively. More testing tomorrow.

Thanks!

Patch-set: 4
Label: Code-Review=0