Update patch set 5
Patch Set 5: Code-Review+1

> Vignesh can explain failures; I haven't looked into them because he knows what's wrong

In all the failures, the gold needs to be updated to account for something we fixed in the library (i.e.) the gold files are wrong. I will follow up with a CL that fixes all the incorrect gold files with clear explanation.

Once again, leaving it to James for the +2.

Patch-set: 5
Reviewer: Vignesh Venkatasubramanian <1002174@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1