Update patch set 1
Patch Set 1:

Not a big deal. Actually this is a little different then the proposed VpxCodecPrivate. The length is set to one byte, but I think we might as well keep it one byte, because it is easier to deal with and I don't think we will ever define any other values.

Patch-set: 1
Reviewer: Frank Galligan <1002054@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0