Update patch set 6
Patch Set 6: Code-Review+1

> This might need some discussion and/or a little extra thought
 > before it lands. I don't see any reason to stick with a default of
 > 2. We have these two bugs:
 > 
 > https://bugs.chromium.org/p/webm/issues/detail?id=1201
 > https://bugs.chromium.org/p/webm/issues/detail?id=1202
 > 
 > Note: both say DocTypeReadVersion, but really mean DocTypeVersion
 > afaik.
 > 
 > I'm pretty sure mkvparser was the only parser silly enough to care
 > about doc type version, and it doesn't anymore. TBH I don't know
 > why we allow users any control of the value, but that ship sailed
 > long ago. Any objections to making 4 the default and not worrying
 > about 1201 and 1202?

This comment made me go back and read the spec for DocTypeVersion and DocTypeReadVersion once again.

It looks like DocTypeVersion just tells the version with which the file was created (i.e.) the max doctype number supported by the writer. So "4" is actually the legit value since our writer can write elements in doctype 4. The parser should completely ignore this value. I don't know why anybody would ever need this value.

DocTypeReadVersion on the other hand is what the parser is supposed to care about when determining its ability to parse the file. We always write that as "2" for this [1]. Both ffmpeg [2] and ExoPlayer [3] would fail if we wrote "4" here (which is what it should be if any of the 4+ elements are present in the file).

So if i am reading the spec correctly, this is broken on multiple levels. I don't see why this change should be a problem.

Will leave it to James to +2 this.

[1] https://chromium.googlesource.com/webm/libwebm/+/master/mkvmuxer/mkvmuxer.cc#127
[2] https://github.com/FFmpeg/FFmpeg/blob/master/libavformat/matroskadec.c#L2410 (even though the variable is called doctype_version, the element it is reading from is DocTypeReadVersion).
[3] https://github.com/google/ExoPlayer/blob/release-v1/library/src/main/java/com/google/android/exoplayer/extractor/webm/WebmExtractor.java#L530

Patch-set: 6