Update patch set 3
Patch Set 3:

The first of... not sure how many of these there are since we don't really know what each file is testing (yet). 

I sent around the sheet where I'll document the migration (for sanity, later). The only functional change (as far as parser behavior is concerned) here is that we die on the call to Cluster::GetFirst(BlockEntry*) instead of after 286 calls to Cluster::GetNext(BlockEntry*, BlockEntry*). Failure occurs on the same line of code; path there is just shorter.

Patch-set: 3
Label: Code-Review=0