Update patch set 4
Patch Set 4: Code-Review+2

You can ignore the clang-format build warning in files not in your CL (the -1 would get over-written when/if the build succeeds).

It looks like it's because of the newer clang-format version in jenkins. Once this lands, i'll do a clang-format pass on all the files with the newer version.

Patch-set: 4
Label: Code-Review=+2