Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Please don't rebase the patch. Like libvpx, you need to land one
>  > patch in order to be able to get past the check in the commit
>  > trigger.
> 
> Is this true for amending to change commit message ?

Yes; the commit message is part of the patch. The review site won't allow me to re-upload a patch without changing it in some way (rebase, change to commit, change to message, etc) so typically I will try to make a very minor/inconsequential change to the commit message.

Now that Tom has made a comment though, you need to address that and upload a new patch. If you put an extra line in the commit message, I can remove that extra line when I reupload the patch.

Sorry for the extra work, but the committer check is not shared across repositories.

Patch-set: 5