Added Type enum in order to support live streams
The mkvparser seems to have been designed with only file based
streams in mind. So it has an issue where it will just continually
load new Clusters, which when streaming a live stream this has the
potential to just go on forever.

Added a Type enum to make it possible to have it load only one Cluster
at a time.